+++ src/database/nedb/chats/messagesRepository.ts
@@ -1,11 +1,14 @@
 import { ZodError } from "zod";
-import { messageDB } from "../initialisation";
+import Datastore from "nedb";
 import { zodParserError } from "@/lib/parser";
-import { InsertMessage, UpdateMessage } from "./model";
+import { InsertMessage, MessageInfer, UpdateMessage } from "./model";
 import { AiMessageSchemaInfer } from "./dto/schema";
+import { initMessageDatastore } from "../initialisation";
 
 export const createMessage = async (inputs: InsertMessage) => {
     try {
+        const messageDB: Datastore<MessageInfer> | undefined = initMessageDatastore();
+        if (!messageDB) return;
         return new Promise((resolve, reject) => {
             messageDB.insert(inputs, (err, doc) => {
                 if (err) reject(err);
@@ -28,6 +31,8 @@ export const getMessagesByConversation = async (
     conversationId: string,
 ): Promise<AiMessageSchemaInfer[] | undefined> => {
     try {
+        const messageDB: Datastore<MessageInfer> | undefined = initMessageDatastore();
+        if (!messageDB) return;
         return new Promise((resolve, reject) => {
             messageDB
                 .find({ conversationId })
@@ -47,6 +52,8 @@ export const getMessagesByConversation = async (
 
 export const updateMessage = async (id: string, updates: UpdateMessage) => {
     try {
+        const messageDB: Datastore<MessageInfer> | undefined = initMessageDatastore();
+        if (!messageDB) return;
         return new Promise((resolve, reject) => {
             messageDB.update(
                 {
@@ -70,6 +77,8 @@ export const updateMessage = async (id: string, updates: UpdateMessage) => {
 
 export const deleteMessage = async (id: string) => {
     try {
+        const messageDB: Datastore<MessageInfer> | undefined = initMessageDatastore();
+        if (!messageDB) return;
         return new Promise((resolve, reject) => {
             messageDB.remove({ _id: id }, (err, doc) => {
                 if (err) reject(err);
@@ -86,6 +95,8 @@ export const deleteMessage = async (id: string) => {
 
 export const getOneMessage = async (id: string) => {
     try {
+        const messageDB: Datastore<MessageInfer> | undefined = initMessageDatastore();
+        if (!messageDB) return;
         return new Promise((resolve, reject) => {
             messageDB.remove({ _id: id }, (err, doc) => {
                 if (err) reject(err);
